+++ docs/CODE_CONVENTION.md
@@ -81,18 +81,22 @@
 
 `FastAPI`를 사용한 웹 애플리케이션 개발 시 다음 원칙을 따릅니다.
 
--   **라우터 분리:**
-    -   기능별로 라우터를 분리하여 `src/web/routers/` 디렉토리에 위치시킵니다.
-    -   각 라우터는 단일 도메인의 엔드포인트만 담당합니다. (예: `stt_router.py`, `template_router.py`)
+-   **엔드포인트 관리:**
+    -   모든 엔드포인트는 `src/web/main.py` 파일에서 중앙 집중식으로 관리합니다.
+    -   기능별로 주석으로 구분하여 가독성을 높입니다. (예: `# ==================== STT & Analysis Endpoints ====================`)
+    -   엔드포인트가 과도하게 많아질 경우에만 라우터 분리를 고려합니다.
 -   **의존성 주입:**
     -   FastAPI의 `Depends`를 활용하여 서비스 클래스를 주입받아 사용합니다.
     -   의존성은 함수 레벨에서 정의하여 재사용성을 높입니다.
+    -   전역 상태가 필요한 경우 `lifespan` 컨텍스트 매니저를 활용합니다.
 -   **응답 모델:**
     -   모든 API 응답은 Pydantic 모델을 사용하여 타입 안정성을 보장합니다.
-    -   응답 모델은 `src/utils/schemas.py`에 정의합니다.
+    -   모든 스키마는 `src/utils/schemas.py`에 통합 관리합니다.
+    -   스트리밍 응답의 경우 `StreamingResponse`를 사용하며 별도 스키마 정의는 불필요합니다.
 -   **예외 처리:**
     -   FastAPI의 `HTTPException`을 사용하여 적절한 HTTP 상태 코드와 메시지를 반환합니다.
-    -   비즈니스 로직에서 발생하는 예외는 라우터 레벨에서 HTTP 예외로 변환합니다.
+    -   비즈니스 로직에서 발생하는 예외는 엔드포인트 레벨에서 HTTP 예외로 변환합니다.
+    -   구체적인 예외 타입을 먼저 처리하고, 일반 `Exception`은 마지막에 처리합니다.
 
 ---
