+++ src/services/template_generator/generate_email.py
@@ -3,7 +3,7 @@ from langchain_core.prompts import ChatPromptTemplate
 
 from src.prompts.template_generation.email_prompts import HUMAN_PROMPT, SYSTEM_PROMPT
 from src.utils.model import llm
-from src.utils.template_schemas import EmailGeneratorInput, EmailGeneratorOutput
+from src.utils.schemas import EmailGeneratorInput, EmailGeneratorOutput
 from src.utils.utils import get_user_data_by_id
 
 def get_email_generator_chain():
